+++ src/data/data_source.rs
@@ -4,9 +4,8 @@ use std::ffi::OsString;
 use std::fmt;
 use std::path::PathBuf;
 
-use braid_codebook::csv::codebook_from_csv;
+use braid_codebook::csv::{codebook_from_csv, ReaderGenerator};
 use braid_codebook::{Codebook, ColMetadataList};
-use csv::ReaderBuilder;
 
 use super::error::DefaultCodebookError;
 
@@ -17,6 +16,8 @@ pub enum DataSource {
     Postgres(PathBuf),
     /// CSV file
     Csv(PathBuf),
+    /// GZipped CSV
+    GzipCsv(PathBuf),
     /// Empty (A void datasource).
     Empty,
 }
@@ -25,7 +26,9 @@ impl TryFrom<DataSource> for PathBuf {
     type Error = &'static str;
     fn try_from(src: DataSource) -> Result<Self, Self::Error> {
         match src {
-            DataSource::Postgres(s) | DataSource::Csv(s) => Ok(s),
+            DataSource::Postgres(s)
+            | DataSource::Csv(s)
+            | DataSource::GzipCsv(s) => Ok(s),
             DataSource::Empty => {
                 Err("DataSource::EMPTY has no path information")
             }
@@ -48,7 +51,9 @@ impl fmt::Display for DataSource {
 impl DataSource {
     pub fn to_os_string(&self) -> Option<OsString> {
         match self {
-            DataSource::Postgres(s) | DataSource::Csv(s) => Some(s),
+            DataSource::Postgres(s)
+            | DataSource::Csv(s)
+            | DataSource::GzipCsv(s) => Some(s),
             DataSource::Empty => None,
         }
         .map(|x| x.clone().into_os_string())
@@ -57,14 +62,16 @@ impl DataSource {
     /// Generate a default `Codebook` from the source data
     pub fn default_codebook(&self) -> Result<Codebook, DefaultCodebookError> {
         match &self {
-            DataSource::Csv(s) => ReaderBuilder::new()
-                .has_headers(true)
-                .from_path(s)
-                .map_err(DefaultCodebookError::CsvError)
-                .and_then(|csv_reader| {
-                    codebook_from_csv(csv_reader, None, None, true)
-                        .map_err(DefaultCodebookError::FromCsvError)
-                }),
+            DataSource::Csv(s) => {
+                let generator = ReaderGenerator::Csv(s.clone());
+                codebook_from_csv(generator, None, None, true)
+                    .map_err(DefaultCodebookError::FromCsvError)
+            }
+            DataSource::GzipCsv(s) => {
+                let generator = ReaderGenerator::GzipCsv(s.clone());
+                codebook_from_csv(generator, None, None, true)
+                    .map_err(DefaultCodebookError::FromCsvError)
+            }
             DataSource::Empty => Ok(Codebook::new(
                 "Empty".to_owned(),
                 ColMetadataList::new(vec![]).unwrap(),