+++ Changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.38.0
+- Source data can now be provided as a GZipped CSV (`.csv.gz`) as well as plain
+    CSV, for both codebook generation and engine running
 - Update Count hyper prior to prevent so many underflow/overflow errors and
     provide better fit
 - Braid should automatically convert old metadata into new metadata